requirements.txt
Click==7.0
Flask==1.0.3
gcloud==0.17.0
googleapis-common-protos==1.6.0
httplib2==0.12.3
hupper==1.6.1
itsdangerous==1.1.0
Jinja2==2.10.1
jws==0.1.3
MarkupSafe==1.1.1
oauth2client==3.0.0
paho-mqtt==1.4.0
PasteDeploy==2.0.1
pkg-resources==0.0.0
plaster==1.0
plaster-pastedeploy==0.7
protobuf==3.8.0
pyasn1==0.4.5
pyasn1-modules==0.2.5
pycryptodome==3.4.3
pyramid==1.10.1
Pyrebase==3.0.27
python-jwt==2.0.1
requests==2.11.1
requests-toolbelt==0.7.0
rsa==4.0
six==1.12.0
translationstring==1.3
venusian==1.2.0
WebOb==1.8.5
Werkzeug==0.15.4
zope.deprecation==4.4.0
zope.interface==4.6.0
